added .env and likes post

+++ app/routers/post.py
@@ -1,4 +1,5 @@
-from typing import List
+from email.headerregistry import Group
+from typing import List, Optional
 from fastapi import (
     Response,
     status,
@@ -7,6 +8,7 @@ from fastapi import (
     APIRouter
 )
 from sqlalchemy.orm import Session
+from sqlalchemy import func
 
 from .. import (
     models, 
@@ -23,16 +25,32 @@ router = APIRouter(
 
 
 
-@router.get("/", response_model=List[schemas.Post])
-def get_posts(db: Session = Depends(get_db)):
-    posts = db.query(models.Post).all()
+@router.get("/", response_model=List[schemas.PostOut])
+def get_posts(db: Session = Depends(get_db),
+              limit: int = 10, 
+              skip: int = 0, 
+              search: Optional[str] = "",
+):
+    posts = (db.query(models.Post, func.count(models.Vote.post_id).label("votes"))
+                .join(models.Vote, models.Vote.post_id == models.Post.id, isouter=True)
+                .group_by(models.Post.id)
+                .filter(models.Post.title.contains(search))
+                .limit(limit)
+                .offset(skip)
+                .all())
 
-    return posts 
+    return posts
 
 
-@router.get("/{id}", response_model=schemas.Post)
+@router.get("/{id}", response_model=schemas.PostOut)
 def get_post(id: int, db: Session = Depends(get_db)):
-    post = db.query(models.Post).filter(models.Post.id == id).first()
+    post = (db.query(models.Post, func.count(models.Vote.post_id)
+              .label("votes"))
+              .join(models.Vote, models.Vote.post_id == models.Post.id, isouter=True)
+              .group_by(models.Post.id)
+              .filter(models.Post.id == id)
+              .first())
+
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"post with id: {id} was not found")
@@ -60,11 +78,16 @@ def delete_post(id: int,
                 
     post_query = db.query(models.Post).filter(models.Post.id == id)
 
-    post = post_query.first() 
-    
+    post = post_query.first()
+
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"post with id: {id} was not found")
+
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, 
+                            detail="Not authorized to perform requested action")
+
     post_query.delete(synchronize_session=False)
     db.commit() 
 
@@ -75,7 +98,7 @@ def delete_post(id: int,
 def update_post(id: int, 
                 updated_post: schemas.CreatePost, 
                 db: Session = Depends(get_db),
-                current_user_id: models.User = Depends(oath2.get_current_user)):
+                current_user: models.User = Depends(oath2.get_current_user)):
 
     post_query = db.query(models.Post).filter(models.Post.id == id)
 
@@ -84,8 +107,12 @@ def update_post(id: int,
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"post with id: {id} was not found")
+
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, 
+                            detail="Not authorized to perform requested action")
     
     post_query.update(updated_post.dict(), synchronize_session=False)
     db.commit()
 
-    return post_query.first()
+    return post
